+++ pkg/pallets/pkg.go
@@ -272,8 +272,9 @@ func (s PkgDeplSpec) ResAttachmentSource(parentSource []string) []string {
 	return append(parentSource, "deployment specification")
 }
 
-// DefinesStack determines whether the PkgDeplSpec instance defines a Docker stack to be deployed.
-func (s PkgDeplSpec) DefinesStack() bool {
+// DefinesApp determines whether the PkgDeplSpec instance defines a Docker Compose app to be
+// deployed.
+func (s PkgDeplSpec) DefinesApp() bool {
 	return s.DefinitionFile != ""
 }